+++ pkg/campaign/controllers/phish.go
@@ -114,11 +114,12 @@ func (ps *PhishingServer) Start() {
 			log.Fatal(err)
 		}
 		log.Infof("Starting phishing server at https://%s", ps.config.ListenURL)
-		log.Fatal(ps.server.ListenAndServeTLS(ps.config.CertPath, ps.config.KeyPath))
+		logServerError(ps.server.ListenAndServeTLS(ps.config.CertPath, ps.config.KeyPath))
+		return
 	}
 	// If TLS isn't configured, just listen on HTTP
 	log.Infof("Starting phishing server at http://%s", ps.config.ListenURL)
-	log.Fatal(ps.server.ListenAndServe())
+	logServerError(ps.server.ListenAndServe())
 }
 
 // Shutdown attempts to gracefully shutdown the server.